requirements.txt
Pillow
numpy
fastapi
uvicorn
typing-extensions
pydantic
tensorflow
keras
python-multipart